click progress bar to move to step
